:recycle: Separate file for DifferenciableIndexPath
