+++ Conv/Algorithm/Algorithm.swift
@@ -8,34 +8,6 @@
 
 import Foundation
 
-struct DifferenciableIndexPath: Differenciable {
-    let uuid: String
-    
-    let section: Section
-    let item: ItemDelegate
-    
-    let sectionIndex: Int
-    let itemIndex: Int
-    
-    var differenceIdentifier: DifferenceIdentifier {
-        return item.differenceIdentifier
-    }
-    
-    func shouldUpdate(to compare: Differenciable) -> Bool {
-        return item.shouldUpdate(to: compare)
-    }
-    
-    var indexPath: IndexPath {
-        return IndexPath(item: itemIndex, section: sectionIndex)
-    }
-}
-
-extension DifferenciableIndexPath: CustomStringConvertible {
-    var description: String {
-        return "section: \(sectionIndex), item: \(itemIndex)"
-    }
-}
-
 struct OperationSet {
     var sectionInsert: [Int] = []
     var sectionUpdate: [Int] = []
